common: move table schema extractor to ddl normalizer
